Fix long-standing bug in MatMPIAIJDiagonalScaleLocalSetUp()
This commit fixes a problem I discovered in October 2017 and then forgot about
after some unsuccessful initial attempts to find the problem. The MPIAIJ
implementation for MatDiagonalScaleLocalSetUp() was broken when the
local-to-global mappings in PETSc were refactored to always use
block indices; the MPIAIJ implementation of
MatDiagonalScaleLocalSetUp(), however, used element-wise, not block
indexing, and this was not refactored at the time, as it was overlooked
due to having no test case.
My commit finally fixes this; it also fixes several memory leaks in the
AIJ and BAIJ implementations of MatDiagonalScaleLocal() which were
overlooked, again, due to having no test cases.